agent-backend/requirements_extraction.txt
# 文件内容提取相关依赖包

# PDF文件处理
PyPDF2>=3.0.0
PyMuPDF>=1.23.0

# Word文档处理
python-docx>=0.8.11
docx2txt>=0.8

# 其他可选依赖
# 如果需要更多格式支持，可以添加以下包：
# openpyxl>=3.1.0  # Excel文件
# python-pptx>=0.6.21  # PowerPoint文件
# python-docx2txt>=0.8  # 另一个DOCX处理库 